make panic filter log better
